+++ CHANGELOG.rst
@@ -10,4 +10,5 @@ Added
 -----
 
 * Model `Predictor` to control inference
-* utility functions `dump_model` and `load_model` to manage saving/loading operations
+* Utility functions `dump_model` and `load_model` to manage saving/loading operations
+* Test for `Predictor` model